bugfix前的修改
