+++ go.mod
@@ -4,7 +4,7 @@ go 1.13
 
 require (
 	github.com/gin-gonic/gin v1.5.0
-	github.com/guonaihong/clop v0.0.3
-	github.com/guonaihong/gout v0.0.11
+	github.com/guonaihong/clop v0.0.7
+	github.com/guonaihong/gout v0.0.12
 	github.com/stretchr/testify v1.4.0
 )
